changed all pngs to svgs for better image quality
