+++ app/layout.js
@@ -6,8 +6,8 @@ import './globals.css'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'youCode 2024',
+  description: 'by WiDS@UBC x UBC WiCS',
 }
 
 export default function RootLayout({ children }) {
