+++ components/Stats.jsx
@@ -19,11 +19,11 @@ const Stats = () => {
       <div className="flex flex-col p-10">
         <div className="flex"> 
           <div className="mx-10 w-1/2 relative" ref={cloud1Ref} style={cloud1Style}>
-            <Image src="/assets/cloud.png" height={150} width={350} layout="responsive" alt="cloud" />
+            <Image src="/assets/cloud3.svg" height={150} width={350} layout="responsive" alt="cloud" />
             <p className='text-3xl font-bold text-dark-green absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-10 font-bold '>100+ Participants</p>
           </div>
           <div className="mx-10 w-1/2 relative" ref={cloud2Ref} style={cloud2Style}>
-            <Image src="/assets/cloud.png" height={150} width={300} layout="responsive" alt="cloud" />
+            <Image src="/assets/cloud4.svg" height={150} width={300} layout="responsive" alt="cloud" />
             <p className='text-3xl font-bold text-dark-green absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-10'>30+ Projects</p>
           </div>
         </div>
