+++ components/Stats2.jsx
@@ -21,12 +21,12 @@ const Stats2 = () => {
       <div className="flex flex-col p-10 ml-40">
         <div className="flex w-2/3"> 
           <div className="mx-20 w-1/2 relative " ref={cloud1Ref} style={cloud1Style}>
-            <Image src="/assets/cloud.png" height={150} width={350} layout="responsive" alt="cloud" />
+            <Image src="/assets/cloud1.svg" height={150} width={350} layout="responsive" alt="cloud" />
             <p className='text-3xl font-bold text-dark-green absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-10'>600+ applications 
 </p>
           </div>
           <div className="mr-16 w-1/2 relative" ref={cloud2Ref} style={cloud2Style}>
-            <Image src="/assets/cloud.png" height={150} width={300} layout="responsive" alt="cloud"/>
+            <Image src="/assets/cloud2.svg" height={150} width={300} layout="responsive" alt="cloud"/>
             <p className='text-3xl font-bold text-dark-green absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 z-10'>25+ Volunteers & Mentors </p>
           </div>
         </div>
