+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@sanity/cli": "^3.21.0",
+    "animejs": "^3.2.2",
     "next": "14.0.3",
     "react": "^18",
     "react-dom": "^18",
