+++ sections/StatsAndAbout.jsx
@@ -2,7 +2,11 @@ import React from 'react'
 import { useParallax } from 'react-scroll-parallax';
 import styled from 'styled-components';
 import { SectionContainer } from '../components/SectionContainer';
+import Stats from '@/components/Stats';
+import About from '@/components/About';
 import TimeLine from '@/components/TimeLine';
+import Stats2 from '@/components/Stats2';
+
 
 const InfoContainer = styled.div`
   background: #150C27;
@@ -18,7 +22,7 @@ const InfoContainer = styled.div`
   
 `
 const BgScroll = styled(SectionContainer)`
-  background: url('assets/Timer3.png');
+  background: url('assets/about_background.svg');
   background-size: 100vw;
   background-repeat: no-repeat;
   background-position: center center;
@@ -30,27 +34,25 @@ const BgScroll = styled(SectionContainer)`
 
 
 `
-
-
-
 const BodyContainer = styled.div`
   
+  margin-top: 2rem;
   position:relative;
+  margin-left: 2rem;
+  margin-right: 2rem;
   
  
 `
 
-
-
-const Timer = () => {
+const StatsAndAboutAndTimeLine = () => {
   return (
     <InfoContainer>
       <BgScroll />
-      <BodyContainer className='flex flex-col items-center justify-center z-10'>
-        
-      <TimeLine/>
+      <BodyContainer className='flex w-full flex-col z-10'>
+     
+      <About/>
+      <Stats2/>
       
-        
       </BodyContainer>
       
       
@@ -58,4 +60,4 @@ const Timer = () => {
   )
 }
 
-export default Timer
+export default StatsAndAboutAndTimeLine
